feat: live per-resource hardware metrics in the dashboard
The metrics panel never showed data: the cpu/ram/network histories were
never populated. Wire them to the selected resource's live statistics,
fetched in the background when the selection changes:

- servers use the current (non-deprecated) statistics endpoint — CPU load
  and network throughput;
- apps use their statistics endpoint — CPU, RAM and network.

The panel titles itself with the selected resource, renders only the
metrics that resource reports (servers expose no live RAM), and shows a
hint on tabs without live statistics. Network samples are stored as f64 to
match the API; the only narrowing is the sparkline's required u64 boundary,
in a scoped helper. No deprecated endpoints, no blanket lint allows.

+++ src/tui/app/tests.rs
@@ -179,8 +179,8 @@ fn push_ram_rolling_window() {
 #[test]
 fn push_net_in_rolling_window() {
     let mut app = App::new(5);
-    for i in 0..65u64 {
-        app.push_net_in(i);
+    for i in 0..65i32 {
+        app.push_net_in(f64::from(i));
     }
     assert_eq!(app.net_in_history.len(), 60);
 }
@@ -188,8 +188,8 @@ fn push_net_in_rolling_window() {
 #[test]
 fn push_net_out_rolling_window() {
     let mut app = App::new(5);
-    for i in 0..65u64 {
-        app.push_net_out(i);
+    for i in 0..65i32 {
+        app.push_net_out(f64::from(i));
     }
     assert_eq!(app.net_out_history.len(), 60);
 }
